Latest changes to public repo to support GAR deployment
