+++ src/main/antlr4/com/pdsl/grammars/AllGrammarsParser.g4
@@ -1,0 +1,5 @@
+parser grammar AllGrammarsParser;
+
+options {tokenVocab=AllGrammarsLexer; }
+
+polymorphicDslAllRules : ALL_INPUTS+;
