+++ src/main/java/com/pdsl/xray/models/XrayTestExecutionResult.java
@@ -16,7 +16,10 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-public record XrayTestExecutionResult(Info info, List<XrayTestResult> tests) {
 
+/**
+ * Wrapper for the DTO for request, to update the Xray API
+ */
+public record XrayTestExecutionResult(Info info, List<XrayTestResult> tests) {
 
 }
