+++ src/main/java/com/pdsl/xray/models/XrayTestResult.java
@@ -14,6 +14,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+/**
+ * DTO for request, to update the Xray API
+ */
 public record XrayTestResult(String testKey, String status) {
 
 }
