+++ src/test/antlr4/com/pdsl/grammars/AllGrammarsLexer.g4
@@ -1,0 +1,5 @@
+lexer grammar AllGrammarsLexer;
+
+// This grammar is used by various tests that don't actually care what the input is
+
+ALL_INPUTS : .* ;
